feat(login): Create a dedicated LoginForm component for token and OAuth authentication
- Introduced a new LoginForm component to handle user authentication via token and OAuth providers (Google and GitHub).
- Removed the previous inline login logic from the page component for better separation of concerns.
- Implemented error handling for invalid tokens and loading states during sign-in.
- Enhanced user experience with a redirect feature upon successful authentication.
