+++ src/app/login/page.tsx
@@ -1,160 +1,5 @@
-"use client";
-
-import { signIn, useSession } from "next-auth/react";
-import { useRouter, useSearchParams } from "next/navigation";
-import { Suspense, useEffect, useState } from "react";
-
-function LoginForm() {
-  const { data: session, status } = useSession();
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const [token, setToken] = useState("");
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-  const callbackUrl = searchParams.get("callbackUrl") || "/";
-
-  useEffect(() => {
-    if (status === "authenticated" && session) {
-      router.push(callbackUrl);
-    }
-  }, [session, status, router, callbackUrl]);
-
-  async function handleTokenLogin(e: React.FormEvent) {
-    e.preventDefault();
-    if (!token.trim()) return;
-
-    setLoading(true);
-    setError(null);
-
-    const result = await signIn("token", {
-      token: token.trim(),
-      redirect: false,
-    });
-
-    if (result?.error) {
-      setError("Invalid token or insufficient permissions. Need an admin-scoped pmth_ token.");
-      setLoading(false);
-    } else if (result?.ok) {
-      router.push(callbackUrl);
-    }
-  }
-
-  if (status === "authenticated") {
-    return (
-      <div className="flex items-center justify-center min-h-screen bg-zinc-950">
-        <div className="animate-pulse text-zinc-500">Redirecting...</div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex items-center justify-center min-h-screen bg-zinc-950">
-      <div className="w-full max-w-sm">
-        <div className="text-center mb-8">
-          <h1 className="text-3xl font-bold tracking-tight">
-            <span className="text-emerald-400">pymt</span>house
-          </h1>
-          <p className="text-zinc-500 mt-2 text-sm">
-            Livepeer Payment Clearinghouse
-          </p>
-        </div>
-
-        {/* Token login */}
-        <div className="border border-zinc-800 rounded-xl p-6 bg-zinc-900/30 mb-4">
-          <h2 className="text-lg font-semibold text-zinc-200 mb-1">
-            Admin Sign In
-          </h2>
-          <p className="text-sm text-zinc-500 mb-5">
-            Paste your admin bearer token to sign in.
-          </p>
-
-          <form onSubmit={handleTokenLogin} className="space-y-3">
-            <div>
-              <input
-                type="password"
-                value={token}
-                onChange={(e) => {
-                  setToken(e.target.value);
-                  setError(null);
-                }}
-                placeholder="pmth_..."
-                className="w-full px-3 py-2.5 bg-zinc-800/50 border border-zinc-700 rounded-lg text-sm text-zinc-200 focus:outline-none focus:border-emerald-500/50 font-mono placeholder:font-sans placeholder:text-zinc-600"
-              />
-            </div>
-
-            {error && (
-              <p className="text-xs text-red-400 bg-red-500/10 border border-red-500/20 rounded-lg px-3 py-2">
-                {error}
-              </p>
-            )}
-
-            <button
-              type="submit"
-              disabled={loading || !token.trim()}
-              className="w-full px-4 py-2.5 bg-emerald-500 text-white rounded-lg text-sm font-medium hover:bg-emerald-600 transition-colors disabled:opacity-50 disabled:cursor-not-allowed"
-            >
-              {loading ? "Signing in..." : "Sign in with Token"}
-            </button>
-          </form>
-
-          <p className="text-xs text-zinc-600 mt-3">
-            Generate a token with: <code className="text-zinc-500">npm run bootstrap</code>
-          </p>
-        </div>
-
-        {/* OAuth providers -- shown when configured */}
-        <div className="border border-zinc-800 rounded-xl p-6 bg-zinc-900/30">
-          <p className="text-xs text-zinc-500 uppercase tracking-wider mb-4">
-            Or sign in with OAuth
-          </p>
-          <div className="space-y-3">
-            <button
-              onClick={() => signIn("google", { callbackUrl })}
-              className="w-full flex items-center justify-center gap-3 px-4 py-2.5 border border-zinc-700 rounded-lg hover:bg-zinc-800/50 transition-colors text-sm font-medium text-zinc-200"
-            >
-              <svg className="w-5 h-5" viewBox="0 0 24 24">
-                <path
-                  d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92a5.06 5.06 0 01-2.2 3.32v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.1z"
-                  fill="#4285F4"
-                />
-                <path
-                  d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                  fill="#34A853"
-                />
-                <path
-                  d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-                  fill="#FBBC05"
-                />
-                <path
-                  d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                  fill="#EA4335"
-                />
-              </svg>
-              Google
-            </button>
-
-            <button
-              onClick={() => signIn("github", { callbackUrl })}
-              className="w-full flex items-center justify-center gap-3 px-4 py-2.5 border border-zinc-700 rounded-lg hover:bg-zinc-800/50 transition-colors text-sm font-medium text-zinc-200"
-            >
-              <svg
-                className="w-5 h-5"
-                fill="currentColor"
-                viewBox="0 0 24 24"
-              >
-                <path d="M12 0c-6.626 0-12 5.373-12 12 0 5.302 3.438 9.8 8.207 11.387.599.111.793-.261.793-.577v-2.234c-3.338.726-4.033-1.416-4.033-1.416-.546-1.387-1.333-1.756-1.333-1.756-1.089-.745.083-.729.083-.729 1.205.084 1.839 1.237 1.839 1.237 1.07 1.834 2.807 1.304 3.492.997.107-.775.418-1.305.762-1.604-2.665-.305-5.467-1.334-5.467-5.931 0-1.311.469-2.381 1.236-3.221-.124-.303-.535-1.524.117-3.176 0 0 1.008-.322 3.301 1.23.957-.266 1.983-.399 3.003-.404 1.02.005 2.047.138 3.006.404 2.291-1.552 3.297-1.23 3.297-1.23.653 1.653.242 2.874.118 3.176.77.84 1.235 1.911 1.235 3.221 0 4.609-2.807 5.624-5.479 5.921.43.372.823 1.102.823 2.222v3.293c0 .319.192.694.801.576 4.765-1.589 8.199-6.086 8.199-11.386 0-6.627-5.373-12-12-12z" />
-              </svg>
-              GitHub
-            </button>
-          </div>
-          <p className="text-xs text-zinc-600 mt-3">
-            Requires OAuth credentials in .env
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-}
+import { Suspense } from "react";
+import { LoginForm } from "./login-form";
 
 export default function LoginPage() {
   return (
